Shell collapse: hoist standalone Plan Viewer tab management to Ui.StandalonePlanViewerController
The standalone Plan Viewer's ~300-line tab-management core (the __PLAN_ADD_TAB__
"+"-tab sentinel, empty "New Plan" sub-tabs, open-file/paste/drag-drop entry, plan
rendering via the shared PlanViewerControl, unique labelling, and teardown) was
duplicated near-verbatim between Lite MainWindow.PlanViewer.cs (425 lines) and the
Darling viewer's (467). It moves to a new PerformanceMonitor.Ui.StandalonePlanViewerController
built with the app's inner plan TabControl; the ForegroundMutedBrush / TabCloseButton
lookups resolve through that control's logical tree, so each app's own theme applies.

Each app keeps only the app-specific pieces: the outer-container reveal/close (Lite
ServerTabControl + select-first-visible vs the viewer's MainTabs + no-servers empty
state), Darling's OpenStoredPlanInPlanViewer store->viewer entry (the FinOps "View
Plan" destination), and the three XAML-wired drag/drop/key handlers as one-line
forwarders.

Faithful extraction: a statement-level equivalence diff (original core, mechanical
renames applied, vs the controller) confirms every prior body statement survives
verbatim; the only controller additions are the ctor + null-guard and Reset() (whose
body is exactly the original close handler's cleanup-loop + clear + init-flag reset).
Net: Lite MainWindow.PlanViewer.cs 425->72, Darling 467->116.

A new PlanViewerCapabilityPinTests ratchet (both test projects, mirroring the ServerTab
pins) matches a probe table against the combined {app xaml + thin app cs + controller}
surface -- 23 probes Lite / 24 Darling -- guarding every plan-loading / drag-drop /
paste / open-file capability and every per-app XAML shell wire against a silent drop.
Each capability was verified present in the pre-hoist source, so the baseline proves
the hoist dropped nothing.

The deprecated Full Dashboard has no standalone Plan Viewer and is untouched. All three
apps build with 0 errors; capability pins green in both Lite (7/7) and Darling (7/7).
