+++ Lite.Tests/CapabilityPins/planviewer.baseline.txt
@@ -1,0 +1,23 @@
+add-tab-sentinel
+clipboard-paste
+ctrl-v-paste
+drag-copy-effect
+drag-filedrop
+empty-state-icon
+empty-state-label
+entry-close
+entry-open
+file-dialog-filter
+multi-file-open
+open-file-button
+open-file-dialog
+paste-button
+render-loadplan
+shared-plan-viewer-control
+unique-sub-tab-label
+viewer-cleanup
+wire-close
+wire-dragover
+wire-drop
+wire-keydown
+wire-open
