+++ PerformanceMonitor.Ui/StandalonePlanViewerController.cs
@@ -1,0 +1,421 @@
+/*
+ * Copyright (c) 2026 Erik Darling, Darling Data LLC
+ *
+ * This file is part of the SQL Server Performance Monitor.
+ *
+ * Licensed under the MIT License. See LICENSE file in the project root for full license information.
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Input;
+
+namespace PerformanceMonitor.Ui;
+
+/// <summary>
+/// Standalone Plan Viewer tab management, shared by Lite's MainWindow and the Darling viewer's MainWindow
+/// (front-end SHELL COLLAPSE - this was ~300 near-verbatim lines duplicated in each app's
+/// <c>MainWindow.PlanViewer.cs</c>). Owns the inner plan <see cref="TabControl"/>'s lifecycle: the "+"
+/// add-tab sentinel, empty "New Plan" sub-tabs and their open-file / paste / drag-drop entry points,
+/// rendering plan XML through the shared <see cref="PlanViewerControl"/>, unique sub-tab labelling, active
+/// sub-tab tracking, and teardown.
+///
+/// The host MainWindow keeps ONLY the app-specific pieces: revealing/hiding the outer server-tabs container
+/// (Lite's <c>ServerTabControl</c> vs the viewer's <c>MainTabs</c>) on open/close, and forwarding its
+/// XAML-wired drag-over / drop / key-down handlers to <see cref="HandleDragOver"/> / <see cref="HandleDrop"/>
+/// / <see cref="HandleKeyDown"/>. Resources (<c>ForegroundMutedBrush</c>, <c>TabCloseButton</c>) resolve
+/// through the injected tab control's logical tree, so each app's own theme dictionaries still apply.
+/// </summary>
+public sealed class StandalonePlanViewerController
+{
+    /* Shared sentinel VALUE - the "+" tab's Tag. Both apps used this exact string; keep it identical so a
+       plan captured with either app's tab layout is indistinguishable. */
+    private const string PlanAddTabId = "__PLAN_ADD_TAB__";
+
+    private readonly TabControl _planTabControl;
+    private bool _initialized;
+
+    public StandalonePlanViewerController(TabControl planTabControl)
+    {
+        _planTabControl = planTabControl ?? throw new ArgumentNullException(nameof(planTabControl));
+    }
+
+    /// <summary>Adds the trailing "+" tab and its selection handler exactly once.</summary>
+    public void EnsureInitialized()
+    {
+        if (_initialized) return;
+        _initialized = true;
+
+        // "+" tab at the end
+        var addTabHeader = new TextBlock
+        {
+            Text = "+",
+            FontSize = 14,
+            FontWeight = FontWeights.Bold,
+            VerticalAlignment = VerticalAlignment.Center,
+            ToolTip = "Open a new plan sub-tab"
+        };
+        var addTab = new TabItem
+        {
+            Header = addTabHeader,
+            Tag = PlanAddTabId,
+            Content = new Grid()
+        };
+        _planTabControl.Items.Add(addTab);
+
+        _planTabControl.SelectionChanged += (_, _) =>
+        {
+            if (_planTabControl.SelectedItem is TabItem { Tag: string t } && t == PlanAddTabId)
+            {
+                var newSub = AddNewEmptyPlanSubTab();
+                _planTabControl.SelectedItem = newSub;
+            }
+        };
+    }
+
+    /// <summary>
+    /// Adds a new empty "New Plan" sub-tab to the inner plan TabControl and selects it.
+    /// Returns the newly created sub-tab.
+    /// </summary>
+    public TabItem AddNewEmptyPlanSubTab()
+    {
+        EnsureInitialized();
+
+        // --- Empty state layer ---
+        var emptyState = new Grid();
+        var dashedRect = new System.Windows.Shapes.Rectangle
+        {
+            Margin = new Thickness(24),
+            Stroke = (System.Windows.Media.Brush)_planTabControl.FindResource("ForegroundMutedBrush"),
+            StrokeThickness = 1.5,
+            StrokeDashArray = new System.Windows.Media.DoubleCollection { 6, 4 },
+            RadiusX = 10, RadiusY = 10,
+            Opacity = 0.25
+        };
+        var emptyStack = new StackPanel { HorizontalAlignment = HorizontalAlignment.Center, VerticalAlignment = VerticalAlignment.Center };
+        emptyStack.Children.Add(new TextBlock
+        {
+            Text = "\uE896",
+            FontFamily = new System.Windows.Media.FontFamily("Segoe MDL2 Assets"),
+            FontSize = 52,
+            HorizontalAlignment = HorizontalAlignment.Center,
+            Foreground = (System.Windows.Media.Brush)_planTabControl.FindResource("ForegroundMutedBrush"),
+            Opacity = 0.45,
+            Margin = new Thickness(0, 0, 0, 12)
+        });
+        emptyStack.Children.Add(new TextBlock
+        {
+            Text = "New Plan",
+            FontSize = 20,
+            FontWeight = FontWeights.Light,
+            Foreground = (System.Windows.Media.Brush)_planTabControl.FindResource("ForegroundMutedBrush"),
+            HorizontalAlignment = HorizontalAlignment.Center
+        });
+        emptyStack.Children.Add(new TextBlock
+        {
+            Text = "Open or paste execution plan XML to render it",
+            FontSize = 13,
+            Foreground = (System.Windows.Media.Brush)_planTabControl.FindResource("ForegroundMutedBrush"),
+            HorizontalAlignment = HorizontalAlignment.Center,
+            Margin = new Thickness(0, 8, 0, 0)
+        });
+        var btnPanel = new StackPanel { Orientation = Orientation.Horizontal, HorizontalAlignment = HorizontalAlignment.Center, Margin = new Thickness(0, 20, 0, 0) };
+        var openBtn = new Button { Content = "Open .sqlplan File", Height = 28, Padding = new Thickness(12, 0, 12, 0), ToolTip = "Open a .sqlplan or .xml file from disk" };
+        var pasteBtn = new Button { Content = "Paste XML", Height = 28, Padding = new Thickness(12, 0, 12, 0), Margin = new Thickness(8, 0, 0, 0), ToolTip = "Paste execution plan XML to render it (or use Ctrl+V)" };
+        btnPanel.Children.Add(openBtn);
+        btnPanel.Children.Add(pasteBtn);
+        emptyStack.Children.Add(btnPanel);
+        emptyStack.Children.Add(new TextBlock
+        {
+            Text = "or drag & drop a .sqlplan file anywhere in this area",
+            FontSize = 11,
+            Foreground = (System.Windows.Media.Brush)_planTabControl.FindResource("ForegroundMutedBrush"),
+            HorizontalAlignment = HorizontalAlignment.Center,
+            Margin = new Thickness(0, 12, 0, 0)
+        });
+        emptyState.Children.Add(dashedRect);
+        emptyState.Children.Add(emptyStack);
+
+        // --- Viewer layer (hidden until a plan is loaded) ---
+        var viewer = new PlanViewerControl
+        {
+            Visibility = Visibility.Collapsed
+        };
+
+        // Sub-tab content grid: index 0 = emptyState, index 1 = viewer
+        var subTabContent = new Grid();
+        subTabContent.Children.Add(emptyState);
+        subTabContent.Children.Add(viewer);
+
+        // --- Sub-tab header: label + close button ---
+        var initialLabel = GetUniqueSubTabLabel("New Plan");
+        var labelBlock = new TextBlock
+        {
+            Text = initialLabel,
+            VerticalAlignment = VerticalAlignment.Center,
+            ToolTip = initialLabel
+        };
+        var subCloseBtn = new Button { Style = (Style)_planTabControl.FindResource("TabCloseButton") };
+        var subTabHeader = new StackPanel { Orientation = Orientation.Horizontal };
+        subTabHeader.Children.Add(labelBlock);
+        subTabHeader.Children.Add(subCloseBtn);
+
+        var subTab = new TabItem { Header = subTabHeader, Content = subTabContent };
+
+        subCloseBtn.Tag = subTab;
+        subCloseBtn.Click += (_, _) =>
+        {
+            (subTabContent.Children[1] as PlanViewerControl)?.Cleanup();
+            _planTabControl.Items.Remove(subTab);
+            // If only the "+" tab remains, open a fresh empty sub-tab
+            if (_planTabControl.Items.Count == 1 &&
+                _planTabControl.Items[0] is TabItem { Tag: string t2 } && t2 == PlanAddTabId)
+            {
+                AddNewEmptyPlanSubTab();
+            }
+        };
+
+        // Wire per-sub-tab buttons
+        openBtn.Click += (_, _) =>
+        {
+            var dialog = new Microsoft.Win32.OpenFileDialog
+            {
+                Filter = "SQL Plan Files (*.sqlplan)|*.sqlplan|XML Files (*.xml)|*.xml|All Files (*.*)|*.*",
+                DefaultExt = ".sqlplan",
+                Multiselect = true
+            };
+            if (dialog.ShowDialog() != true) return;
+            var isFirst = true;
+            foreach (var fileName in dialog.FileNames)
+            {
+                try
+                {
+                    var xml = System.IO.File.ReadAllText(fileName);
+                    var targetTab = isFirst ? subTab : AddNewEmptyPlanSubTab();
+                    LoadPlanIntoSubTab(targetTab, xml, System.IO.Path.GetFileName(fileName));
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show($"Failed to open file:\n\n{ex.Message}", "Error",
+                        MessageBoxButton.OK, MessageBoxImage.Error);
+                }
+                isFirst = false;
+            }
+        };
+
+        pasteBtn.Click += (_, _) =>
+        {
+            var xml = Clipboard.GetText();
+            if (!string.IsNullOrWhiteSpace(xml))
+            {
+                LoadPlanIntoSubTab(subTab, xml, "Pasted Plan");
+                return;
+            }
+            MessageBox.Show("The clipboard does not contain any text.", "Paste Plan XML",
+                MessageBoxButton.OK, MessageBoxImage.Information);
+        };
+
+        // Insert before "+" tab
+        var addTabIndex = -1;
+        for (var i = 0; i < _planTabControl.Items.Count; i++)
+        {
+            if (_planTabControl.Items[i] is TabItem { Tag: string t3 } && t3 == PlanAddTabId)
+            {
+                addTabIndex = i;
+                break;
+            }
+        }
+        if (addTabIndex >= 0)
+            _planTabControl.Items.Insert(addTabIndex, subTab);
+        else
+            _planTabControl.Items.Add(subTab);
+
+        _planTabControl.SelectedItem = subTab;
+        return subTab;
+    }
+
+    /// <summary>
+    /// Loads plan XML into an existing sub-tab (replacing whatever was there before).
+    /// </summary>
+    public async void LoadPlanIntoSubTab(TabItem subTab, string planXml, string label, string? queryText = null)
+    {
+        if (subTab.Content is not Grid subTabContent) return;
+        if (subTabContent.Children.Count < 2) return;
+
+        var emptyState = subTabContent.Children[0] as FrameworkElement;
+        var viewer = subTabContent.Children[1] as PlanViewerControl;
+        if (viewer == null) return;
+
+        try
+        {
+            /* LoadPlan parses+analyzes off the UI thread; XmlException replaces the redundant
+               up-front XDocument.Parse validation. */
+            await viewer.LoadPlan(planXml, label, queryText);
+        }
+        catch (System.Xml.XmlException ex)
+        {
+            MessageBox.Show(
+                $"The plan XML is not valid:\n\n{ex.Message}",
+                "Invalid Plan XML",
+                MessageBoxButton.OK,
+                MessageBoxImage.Warning);
+            return;
+        }
+        catch (Exception ex)
+        {
+            MessageBox.Show(
+                $"Failed to load the execution plan:\n\n{ex.Message}",
+                "Plan Load Error",
+                MessageBoxButton.OK,
+                MessageBoxImage.Warning);
+            return;
+        }
+        emptyState!.Visibility = Visibility.Collapsed;
+        viewer.Visibility = Visibility.Visible;
+
+        var uniqueLabel = GetUniqueSubTabLabel(label);
+        if (subTab.Header is StackPanel headerPanel &&
+            headerPanel.Children[0] is TextBlock headerLabel)
+        {
+            headerLabel.Text = uniqueLabel.Length > 30 ? uniqueLabel[..30] + "\u2026" : uniqueLabel;
+            headerLabel.ToolTip = uniqueLabel;
+        }
+    }
+
+    /// <summary>
+    /// Returns a label that is unique among current inner plan sub-tab headers.
+    /// If <paramref name="baseLabel"/> is already taken, appends " (1)", " (2)", \u2026
+    /// </summary>
+    private string GetUniqueSubTabLabel(string baseLabel)
+    {
+        var existing = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+        foreach (var item in _planTabControl.Items)
+        {
+            if (item is TabItem { Tag: string t } && t == PlanAddTabId) continue;
+            if (item is TabItem subTab &&
+                subTab.Header is StackPanel sp &&
+                sp.Children[0] is TextBlock tb)
+                existing.Add(tb.ToolTip as string ?? tb.Text);
+        }
+        if (!existing.Contains(baseLabel)) return baseLabel;
+        var counter = 1;
+        string candidate;
+        do { candidate = $"{baseLabel} ({counter++})"; }
+        while (existing.Contains(candidate));
+        return candidate;
+    }
+
+    /// <summary>
+    /// Returns the currently active real plan sub-tab (skips the "+" tab).
+    /// </summary>
+    private TabItem? GetActivePlanSubTab()
+    {
+        if (_planTabControl.SelectedItem is TabItem { Tag: string t } && t == PlanAddTabId)
+            return null;
+        return _planTabControl.SelectedItem as TabItem;
+    }
+
+    /// <summary>Drag-over handler (forwarded from the host's XAML wire): accept only plan files.</summary>
+    public void HandleDragOver(DragEventArgs e)
+    {
+        if (e.Data.GetDataPresent(DataFormats.FileDrop))
+        {
+            var files = e.Data.GetData(DataFormats.FileDrop) as string[];
+            if (files?.Any(IsPlanFile) == true)
+            {
+                e.Effects = DragDropEffects.Copy;
+                e.Handled = true;
+                return;
+            }
+        }
+        e.Effects = DragDropEffects.None;
+        e.Handled = true;
+    }
+
+    /// <summary>Drop handler (forwarded from the host's XAML wire): first file into the active tab, rest into new tabs.</summary>
+    public void HandleDrop(DragEventArgs e)
+    {
+        if (!e.Data.GetDataPresent(DataFormats.FileDrop)) return;
+        var planFiles = (e.Data.GetData(DataFormats.FileDrop) as string[])
+            ?.Where(IsPlanFile).ToArray();
+        if (planFiles == null || planFiles.Length == 0) return;
+        LoadPlanFromFileIntoActiveTab(planFiles[0]);
+        for (var i = 1; i < planFiles.Length; i++)
+        {
+            var newTab = AddNewEmptyPlanSubTab();
+            try
+            {
+                var xml = System.IO.File.ReadAllText(planFiles[i]);
+                LoadPlanIntoSubTab(newTab, xml, System.IO.Path.GetFileName(planFiles[i]));
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show($"Failed to load plan file:\n{ex.Message}", "Load Error",
+                    MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+        }
+    }
+
+    /// <summary>Key-down handler (forwarded from the host's XAML wire): Ctrl+V pastes plan XML into the active tab.</summary>
+    public void HandleKeyDown(KeyEventArgs e)
+    {
+        if (e.Key == Key.V &&
+            Keyboard.Modifiers == ModifierKeys.Control &&
+            e.OriginalSource is not TextBox)
+        {
+            var xml = Clipboard.GetText();
+            if (!string.IsNullOrWhiteSpace(xml))
+            {
+                e.Handled = true;
+                LoadPlanIntoActivePlanSubTab(xml, "Pasted Plan");
+            }
+        }
+    }
+
+    private void LoadPlanFromFileIntoActiveTab(string path)
+    {
+        try
+        {
+            var xml = System.IO.File.ReadAllText(path);
+            LoadPlanIntoActivePlanSubTab(xml, System.IO.Path.GetFileName(path));
+        }
+        catch (Exception ex)
+        {
+            MessageBox.Show($"Failed to load plan file:\n{ex.Message}", "Load Error",
+                MessageBoxButton.OK, MessageBoxImage.Error);
+        }
+    }
+
+    private void LoadPlanIntoActivePlanSubTab(string planXml, string label)
+    {
+        var activeSubTab = GetActivePlanSubTab();
+        if (activeSubTab != null)
+            LoadPlanIntoSubTab(activeSubTab, planXml, label);
+    }
+
+    /// <summary>
+    /// Tears the plan surface down to a fresh state: cleans up every sub-tab's viewer (each is rooted by the
+    /// static ThemeChanged event, so must be released before the tab control is cleared), empties the tab
+    /// control, and re-arms <see cref="EnsureInitialized"/> so the next open rebuilds the "+" tab. The host's
+    /// close handler calls this, then restores its own outer container.
+    /// </summary>
+    public void Reset()
+    {
+        foreach (var item in _planTabControl.Items)
+            if (item is TabItem { Content: Grid g } && g.Children.Count > 1
+                && g.Children[1] is PlanViewerControl pv)
+                pv.Cleanup();
+        _planTabControl.Items.Clear();
+        _initialized = false;
+    }
+
+    private static bool IsPlanFile(string path)
+    {
+        var ext = System.IO.Path.GetExtension(path);
+        return string.Equals(ext, ".sqlplan", StringComparison.OrdinalIgnoreCase)
+            || string.Equals(ext, ".xml", StringComparison.OrdinalIgnoreCase);
+    }
+}
